Commits: 1. Software installation package 2. IP/port number input box verification 3. The configured serial port cannot be configured again, and the configured device cannot be added to the site. 4. Add four mobile operation buttons to the image editing function of task management 5. If there is no product picture or any point on a certain product in the task, it cannot be selected on the workbench. 6. Data cannot be stored in selected fields 7. The data document name does not support customization. 8. Data query does not support export 9. The date query input box of data query does not add a date component. 10. If the network is disconnected and then reconnected, the task cannot continue to be executed. 11. The error message when tightening is incorrect is not displayed in the status 12. Add device connection fact log to event log 13. Add quick delete function to the point editing button on the product surface and right side 14. The enumeration items in the data are displayed as Name values 15. The IP in the tool IP field in the data must also include the port number

+++ OperationGuidance_service/Models/DTOs/OperationDataDTO.cs
@@ -52,10 +52,10 @@ namespace OperationGuidance_service.Models.DTOs {
         public int? rundown_angle_max { get; set; }                                         // 贴合角度上限
         public int? rundown_angle { get; set; }                                             // 贴合角度
         public int? rundown_angle_target { get; set; } = 0;                                 // * 贴合角度目标值
-        public int? rundown_torque_min { get; set; } = 0;                                   // * 贴合扭力下限
-        public int? rundown_torque_max { get; set; } = 0;                                   // * 贴合扭力上限
-        public int? rundown_torque { get; set; } = 0;                                       // * 贴合扭力
-        public int? rundown_torque_target { get; set; } = 0;                                // * 贴合扭力目标值
+        public float? rundown_torque_min { get; set; } = 0;                                   // * 贴合扭力下限
+        public float? rundown_torque_max { get; set; } = 0;                                   // * 贴合扭力上限
+        public float? rundown_torque { get; set; } = 0;                                       // * 贴合扭力
+        public float? rundown_torque_target { get; set; } = 0;                                // * 贴合扭力目标值
         public int? current_monitoring_min { get; set; }                                    //
         public int? current_monitoring_max { get; set; }                                    //
         public int? current_monitoring_value { get; set; }                                  //
